test: move keypair tests out of common.rs into tests/generate_keypair.rs
Per review: common.rs is the wrong home for a unit-test module. The
ML-DSA-65 / ed25519 key-generation tests move to an integration test at
tests/generate_keypair.rs, exercising the public API
(near_cli_rs::common::{GeneratedKeyPair, SignatureScheme}).
